Separate Signal.Validation into 2 modules
