+++ lib/sanbase/signals/trigger/settings/daily_active_addresses_settings.ex
@@ -6,7 +6,8 @@ defmodule Sanbase.Signal.Trigger.DailyActiveAddressesSettings do
   """
   use Vex.Struct
 
-  import Sanbase.Signal.{Validation, Utils}
+  import Sanbase.{Validation, Signal.Validation}
+  import Sanbase.Signal.Utils
 
   alias __MODULE__
   alias Sanbase.Signal.Type
