+++ lib/sanbase/signals/trigger/validation.ex
@@ -1,12 +1,7 @@
 defmodule Sanbase.Signal.Validation do
+  import Sanbase.Validation
+
   @notification_channels ["telegram", "email"]
-
-  defguard is_valid_price(price) when is_number(price) and price >= 0
-  defguard is_valid_percent(percent) when is_number(percent) and percent >= -100
-  defguard is_valid_percent_change(percent) when is_number(percent) and percent > 0
-
-  defguard is_valid_min_max_price(min, max)
-           when min < max and is_valid_price(min) and is_valid_price(max)
 
   def valid_notification_channels(), do: @notification_channels
 
@@ -14,11 +9,6 @@ defmodule Sanbase.Signal.Validation do
 
   def valid_notification_channel?(channel),
     do: {:error, "#{inspect(channel)} is not a valid notification channel"}
-
-  def valid_percent?(percent) when is_valid_percent(percent), do: :ok
-
-  def valid_percent?(percent),
-    do: {:error, "#{inspect(percent)} is not a valid percent"}
 
   def valid_percent_change_operation?(%{percent_up: percent})
       when is_valid_percent_change(percent) do
@@ -51,29 +41,6 @@ defmodule Sanbase.Signal.Validation do
 
   def valid_absolute_change_operation?(operation),
     do: {:error, "#{inspect(operation)} is not a valid absolute change operation"}
-
-  def valid_time_window?(time_window) when is_binary(time_window) do
-    Regex.match?(~r/^\d+[smhdw]$/, time_window)
-    |> case do
-      true -> :ok
-      false -> {:error, "#{inspect(time_window)} is not a valid time window"}
-    end
-  end
-
-  def valid_time_window?(time_window),
-    do: {:error, "#{inspect(time_window)} is not a valid time window"}
-
-  def valid_iso8601_datetime_string?(time) when is_binary(time) do
-    case Time.from_iso8601(time) do
-      {:ok, _time} ->
-        :ok
-
-      _ ->
-        {:error, "#{time} isn't a valid ISO8601 time"}
-    end
-  end
-
-  def valid_iso8601_datetime_string?(_), do: {:error, "Not valid ISO8601 time"}
 
   def valid_target?("default"), do: :ok
 
@@ -123,21 +90,6 @@ defmodule Sanbase.Signal.Validation do
   def valid_slug?(slug) do
     {:error,
      "#{inspect(slug)} is not a valid slug. A valid slug is a map with a single slug key and string value"}
-  end
-
-  def valid_url?(url) do
-    case URI.parse(url) do
-      %URI{scheme: nil} -> {:error, "`#{url}` is missing scheme"}
-      %URI{host: nil} -> {:error, "`#{url}` is missing host"}
-      %URI{path: nil} -> {:error, "`#{url}` is missing path"}
-      _ -> :ok
-    end
-  end
-
-  def valid_threshold?(t) when is_number(t) and t > 0, do: :ok
-
-  def valid_threshold?(t) do
-    {:error, "#{inspect(t)} is not valid threshold. It must be a number bigger than 0"}
   end
 
   # private functions
